Make initial commit that implements basic working game
Adds a board hardcoded to 3x3, the ability to place a mark and a
function for checking the win condition, as well as the main game loop
which exits when a player wins.

Everything is implemented as basically as possible so that complex
language structure is avoided but can be added to simplify the source code.